+++ src/app/locales/en/components/-storage/generic.js
@@ -1,4 +1,11 @@
 export default {
-  name: 'Storage name',
-  lumaEnabled: 'LUMA enabled',
+  name: { name: 'Storage name' },
+  lumaEnabled: {
+    name: 'LUMA enabled',
+    tip: 'LUMA allows to map onedata user credentials into storage credentials' +
+      ' and vice versa. If enabled, provided LUMA service will be used to resolve' +
+      ' the mappings during operations on storage. If disabled, some random' +
+      ' credentials (e.g. uid and gid on POSIX storage) will be generated for' +
+      ' every user.',
+  },
 };